refactor: EmailAuth, EmailForm, SignUpForm handler state hook으로 분리
