+++ src/home/components/SignupContents/EmailAuth/EmailAuth.type.ts
@@ -1,0 +1,4 @@
+export interface EmailAuthProps {
+  onConfirm: () => void;
+  email: string;
+}
