+++ src/home/components/SignupContents/EmailForm/EmailForm.tsx
@@ -1,8 +1,7 @@
-import { useState } from 'react';
-
 import { BoxButton, PlainButton, SimpleTextField } from '@yourssu/design-system-react';
 
-import { postAuthVerificationEmail } from '@/home/apis/authVerification.ts';
+import { EmailFormProps } from '@/home/components/SignupContents/EmailForm/EmailForm.type.ts';
+import { useEmailForm } from '@/home/components/SignupContents/EmailForm/useEmailForm.ts';
 
 import {
   StyledSignupButtonText,
@@ -18,30 +17,10 @@ import {
   StyledTextFieldLabel,
 } from './EmailForm.style';
 
-interface EmailFormProps {
-  onConfirm: (email: string) => void;
-}
-
 const EMAIL_DOMAIN = '@soongsil.ac.kr';
 
 export const EmailForm = ({ onConfirm }: EmailFormProps) => {
-  const [email, setEmail] = useState('');
-  const [emailError, setEmailError] = useState<string | undefined>(undefined);
-  const [emailSending, setEmailSending] = useState(false);
-
-  const onChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setEmail(e.target.value);
-  };
-
-  const onEmailSubmit = async () => {
-    let fullEmail = email;
-    if (!email.endsWith(EMAIL_DOMAIN)) fullEmail += EMAIL_DOMAIN;
-    setEmailSending(true);
-    const res = await postAuthVerificationEmail({ email: fullEmail, verificationType: 'SIGN_UP' });
-    if (res.data) onConfirm(fullEmail);
-    else setEmailError(res.error?.response?.data.message || '이메일을 다시 확인해주세요.');
-    setEmailSending(false);
-  };
+  const { email, emailSending, emailError, onEmailSubmit, onChange } = useEmailForm({ onConfirm });
 
   return (
     <StyledSignupContentContainer>
