+++ src/home/components/SignupContents/EmailForm/EmailForm.type.ts
@@ -1,0 +1,3 @@
+export interface EmailFormProps {
+  onConfirm: (email: string) => void;
+}
