+++ src/home/components/SignupContents/SignupForm/SignUpForm.type.ts
@@ -1,0 +1,4 @@
+export interface SignupFormProps {
+  onConfirm: () => void;
+  email: string;
+}
